Muestreo/requirements.txt
pyserial
matplotlib
tk
pandas
openpyxl